ajuste simples na tipagem digitada incorretamente no DTO de ranking

+++ src/main/java/infnet/tp3_springboot/dto/RankingAventureiroDTO.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 public class RankingAventureiroDTO {
     private String nome;
     private Long totalParticipacoes;
-    private Double somaRecompensas;
+    private Long somaRecompensas;
     private Long totalMvps;
 }
